refactor: 구동 리스트 컴포넌트 분리
