+++ front-end/src/pages/Write.jsx
@@ -9,8 +9,8 @@ import { Editor } from "@toast-ui/react-editor";
 
 import "@toast-ui/editor/dist/toastui-editor.css";
 
-import { guList, dongList } from "../data/SeoulDistricts";
 import uploadImages from "../utils/upLoadImages";
+import SelectGuDong from "../components/SelectGuDong";
 import { postBoard } from "../api/board";
 
 const EntireContainer = styled.div`
@@ -52,26 +52,6 @@ const FormSection = styled.form`
 
 const EditorContainer = styled.div`
   margin-bottom: 8%;
-`;
-
-const TagContainer = styled.div`
-  margin-top: 2%;
-  margin-bottom: 5%;
-  > select {
-    height: 30px;
-    width: 78px;
-    margin-right: 10px;
-    border-radius: 20px;
-    padding: 0.5rem;
-    color: #fff;
-    font-weight: 500;
-    background-color: var(--primary-color);
-    cursor: pointer;
-    &:focus {
-      border: 1px solid var(--font-color-bold);
-      outline: 4px solid #f7eaea;
-    }
-  }
 `;
 
 const ButtonContainer = styled.div`
@@ -255,47 +235,5 @@ export default function Write() {
         </button>
       </ButtonContainer>
     </FormSection>
-  );
-}
-
-// 공통 컴포넌트 따로 빼서 리팩토링할 예정
-function SelectGuDong({ onGuChange, onDongChange }) {
-  const [selectedGu, setSelectedGu] = useState("");
-  const [selectedDong, setSelectedDong] = useState("");
-
-  function handleGuChange(event) {
-    setSelectedGu(event.target.value);
-    onGuChange(event.target.value);
-  }
-
-  function handleDongChange(event) {
-    setSelectedDong(event.target.value);
-    onDongChange(event.target.value);
-  }
-
-  return (
-    <TagContainer>
-      <select value={selectedGu} onChange={handleGuChange} required>
-        <option value="" hidden>
-          지역구
-        </option>
-        {guList.map(gu => (
-          <option key={gu} value={gu}>
-            {gu}
-          </option>
-        ))}
-      </select>
-      <select value={selectedDong} onChange={handleDongChange} disabled={!selectedGu} required>
-        <option value="" hidden>
-          지역동
-        </option>
-        {selectedGu &&
-          dongList[selectedGu].map(dong => (
-            <option key={dong} value={dong}>
-              {dong}
-            </option>
-          ))}
-      </select>
-    </TagContainer>
   );
 }
